NVRAM variable key lifetime (reference from the caller) should not be tied to the raw NVRAM buffer.
While all the parsed content of NVRAM can refer to the original data buffer,
it is not the case when we start adding new variables. Both key and value should be Cow.

This could be solved by making key: Cow<'a, [u8]> in the API but that would be a breaking change.